Add cast to support 64 bit.
http://code.google.com/p/skia/issues/detail?id=524
Review URL: https://codereview.appspot.com/6038053

git-svn-id: http://skia.googlecode.com/svn/trunk@3701 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/core/SkFontHost.h
@@ -17,7 +17,7 @@ class SkDescriptor;
 class SkStream;
 class SkWStream;
 
-typedef uintptr_t SkFontTableTag;
+typedef uint32_t SkFontTableTag;
 
 /** \class SkFontHost
